Authenticate and Authorize using JWT

+++ controllers/api/v1/post_api.js
@@ -24,13 +24,20 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        post.remove();
+        if (post.user == req.user.id){
+            post.remove();
 
-        await Comment.deleteMany({post: req.params.id});
+            await Comment.deleteMany({post: req.params.id});
 
-        return res.json(200, {
-            message: "Post and associated comments deleted successfully!!"
-        });
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!!"
+            });
+            
+        }else{
+            return res.json(401, {
+                message: "You cant delete the post !! :("
+            });
+        }
 
     }catch(err){
         console.log("**********"+err);
